data stuff should be done, now need to do user auth

+++ bandform-backend/src/main/java/xyz/wmmp/bandform_backend/services/BandMemberService.java
@@ -20,33 +20,59 @@ public class BandMemberService {
     private final UserRepository userRepository;
     private final BandRepository bandRepository;
     private final InstrumentService instrumentService;
+    private final BandService bandService;
 
     @Autowired
-    public BandMemberService(BandMemberRepository bandMemberRepository, UserRepository userRepository, BandRepository bandRepository, InstrumentService instrumentService){
+    public BandMemberService(BandMemberRepository bandMemberRepository, UserRepository userRepository, BandRepository bandRepository, InstrumentService instrumentService, BandService bandService){
         this.bandMemberRepository = bandMemberRepository;
         this.userRepository = userRepository;
         this.bandRepository = bandRepository;
         this.instrumentService = instrumentService;
+        this.bandService = bandService;
     }
 
     public List<BandMember> membersInBand(Long bandId){
         return bandMemberRepository.findByBandNameIn(bandId).orElse(null);
     }
 
+    public BandMember createBandMember(Band band, User user, List<Instrument> instruments, String role){
+        BandMember bm = new BandMember();
+        bm.setUser(user);
+        bm.setBand(band);
+        bm.setInstruments(instruments);
+        bm.setJoinedDate(LocalDateTime.now());
+        bm.setRole(role);
+        bm = bandMemberRepository.save(bm);
+        List<BandMember> toUpdate = band.getMembers();
+        toUpdate.add(bm);
+        bandService.updateBand(band.getId(), null, null, null, null, null, toUpdate, null, null);
+        return bm;
+    }
+
     public BandMember createBandMember(Long bandId, Long userId, List<String> instrumentNames, String role){
         BandMember bm = new BandMember();
         User u = userRepository.findById(userId).orElse(null);
         bm.setUser(u);
-        Band b = bandRepository.findById(bandId).orElse(null);
+        Band b = bandService.getBandById(bandId);
         bm.setBand(b);
         List<Instrument> i = instrumentService.getInstrumentsByNameAndAddIfNecessary(instrumentNames);
         bm.setInstruments(i);
         bm.setJoinedDate(LocalDateTime.now());
         bm.setRole(role);
-        return bandMemberRepository.save(bm);
+        bm = bandMemberRepository.save(bm);
+        List<BandMember> toUpdate = b.getMembers();
+        toUpdate.add(bm);
+        bandService.updateBand(b.getId(), null, null, null, null, null, toUpdate, null, null);
+        return bm;
     }
 
     public Long removeMember(Long memberId){
+        BandMember bm = bandMemberRepository.findById(memberId).orElse(null);
+        if(bm == null){return null;}
+        Band b = bm.getBand();
+        List<BandMember> toUpdate = b.getMembers();
+        toUpdate.remove(bm);
+        bandService.updateBand(b.getId(), null, null, null, null, null, toUpdate, null, null);
         bandMemberRepository.deleteById(memberId);
         return memberId;
     }
